move logic to models
