+++ app/controllers/matchings_controller.rb
@@ -22,6 +22,15 @@ class MatchingsController < ApplicationController
     end
   end
 
+  def all_submitted_preferences?
+    @all_participants_ids.each do |participant_id|
+      unless Participant.find_by(id: participant_id).last_responded
+        return false
+      end
+    end
+    true
+  end
+
   def modify
     @matching = Matching.find_by(params.slice(:project_id))
     @parsed_matching = JSON.parse(@matching.output_json)
@@ -67,49 +76,20 @@ class MatchingsController < ApplicationController
     flash[:success] = 'Emails have been sent.'
     redirect_to project_matching_path(params[:project_id])
   end
-
-  # GET /projects/:project_id/matching/new
-  # def new
-  #   @matching = Matching.new
-  # end
-
-  # TODO: convert to a PUT/ UPDATE
-  # GET /projects/:project_id/matching/edit
-
-  # def edit
-  #   respond_to do |format|
-  #     if all_submitted_preferences?
-  #       #matching = Matching.where(project_id: @project.id).update_all(output_json: api)
-  #       if @matching.update(output_json: api)
-  #         flash[:success] = 'Successfully matched.'
-  #         format.html { redirect_to project_matching_path }
-  #       # else
-  #       #   format.html { render :edit }
-  #       end
-  #     end
-  #   end
-  # end
-
   # POST /projects/:project_id/matching
   def create
     @project = Project.find_by(id: params[:project_id])
     @matching = @project.matching || Matching.new(project_id: @project.id)
-    @matching.output_json = api
+    @matching.output_json = @matching.api_call
 
     respond_to do |format|
-      allmatched = true
-      notmatched = ""
       if @matching.save!
-        for participant in @project.participants
-          if !@matching.output_json.include? participant.email
-            allmatched = false
-            notmatched += participant.email + ", "
-          end
-        end
+        notmatched = @matching.participants_not_matched
+        allmatched = (notmatched == "")
         if allmatched
           flash[:success] = 'Matching Complete. All users successfully matched.'
         else
-          flash[:success] = 'Matching Complete. ' + notmatched[0...-2] + ' did not receive a match.'
+          flash[:success] = 'Matching Complete. ' + notmatched + ' did not receive a match.'
         end
         format.html { redirect_to project_matching_path }
         # else
@@ -118,91 +98,15 @@ class MatchingsController < ApplicationController
     end
   end
 
-  # DELETE /projects/:project_id/matching
-  # def destroy
-  #   @matching.destroy
-  #   respond_to do |format|
-  #     flash[:success] = 'Matching was successfully destroyed.'
-  #     format.html { redirect_to matchings_url }
-  #   end
-  # end
-
-  def category
-    'person_to_time'
-  end
-
   def global_settings
     { "minutes": @project.duration }
   end
 
-  def people
-    people = []
-    @project.participants.each do |participant|
-      row = { "name": participant.email, "match_degree": participant.match_degree }
-      people.push(row)
-    end
-
-    people
-  end
-
-  def timeslots
-    all_project_times = ProjectTime.where(project_id: @project.id).pluck(:date_time)
-
-    timeslots = []
-    all_project_times.each do |timeslot|
-      timeslot_formatted = timeslot.strftime('%Y-%m-%d %H:%M')
-      row = { "timestamp": timeslot_formatted, "tracks": 1 }
-      timeslots.push(row)
-    end
-
-    timeslots
-  end
-
-  def preferences
-    preferences = []
-    @project.rankings.each do |ranking|
-      person = ranking.participant.email
-      timeslot = ranking.project_time.date_time
-      timeslot_formatted = timeslot.strftime('%Y-%m-%d %H:%M')
-      if ranking.rank != 0
-        row = {"person_name": person, "timeslot": timeslot_formatted, "rank": ranking.rank}
-        preferences.push(row)
-      end
-    end
-
-    preferences
-  end
-
-  def all_submitted_preferences?
-    @all_participants_ids.each do |participant_id|
-      unless Participant.find_by(id: participant_id).last_responded
-        return false
-      end
-    end
-    true
-  end
-
-  def api
-    require 'rest-client'
-
-    input = {
-      category: category,
-      global_settings: global_settings,
-      people: people,
-      timeslots: timeslots,
-      preferences: preferences
-    }
-
-    output = RestClient.post('http://api.multi-meet.com:5000/multimatch', input.to_json,
-                             content_type: :json, accept: :json).body
-    output
-  end
+  
 
   private
     # Use callbacks to share common setup or constraints between actions.
     def set_instance_variables
-      @matching = Matching.find_by(params.slice(:project_id))
-      @project = Project.find_by(:id => params[:project_id])
       set_matching_and_project
       if @project
         @times = @project.project_times.order(:date_time)
